Add font-tables::offset module
This is a more sane place for putting code related to resolving offsets,
and is also where code related to handling Null offsets will likely
live.
